dom/telephony/test/test_dial1.html

<!--
  Any copyright is dedicated to the Public Domain.
  http://creativecommons.org/publicdomain/zero/1.0/
-->
<html>
<head>
  <title>Test for DOM Telephony</title>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<pre id="test">
<script class="testbody" type="text/javascript">

  SimpleTest.waitForExplicitFinish();

  SpecialPowers.setCharPref("dom.telephony.app.phone.url",
                            window.location.href);

  var telephony = window.navigator.mozTelephony;

  isnot(telephony, null, "Have a telephony object");
  ok(telephony === window.navigator.mozTelephony, "Same telephony object");

  var session = telephony.connect("5551212");
  is(session.number, "5551212", "Correct number");
  is(session.readyState, "initial", "Correct state");

  session.ondialing = function(event) {
    ok(event.target === session, "Correct event target");
    is(session.readyState, "dialing", "Correct readyState");
    is(telephony.liveSessions.length, 1, "Correct number of live sessions");
    ok(telephony.liveSessions[0] === session, "Correct session array");
    event.target.hangUp();
  };

  session.ondisconnected = function(event) {
    ok(event.target === session, "Correct event target");
    is(telephony.liveSessions.length, 0, "Correct number of live sessions");
    is(session.readyState, "disconnected", "Correct readyState");
    SimpleTest.finish();
  }

</script>
</pre>
</body>
</html>
